+++ src/Phan/Language/Element/Parameter.php
@@ -28,6 +28,7 @@ use Phan\Library\StringUtil;
 use Phan\Parse\ParseVisitor;
 
 use function is_string;
+use function preg_match;
 use function strlen;
 
 /**
@@ -76,6 +77,12 @@ class Parameter extends Variable
      * True if the default value was inferred from reflection
      */
     private $default_value_from_reflection = false;
+
+    /**
+     * @var bool
+     * True if the variable name or comment indicates the parameter is unused
+     */
+    private $should_warn_if_provided = false;
 
     /**
      * @return static
@@ -304,9 +311,10 @@ class Parameter extends Variable
         }
 
         // Create the skeleton parameter from what we know so far
+        $parameter_name = (string)$node->children['name'];
         $parameter = Parameter::create(
             (clone($context))->withLineNumberStart($node->lineno),
-            (string)$node->children['name'],
+            $parameter_name,
             $union_type,
             $node->flags
         );
@@ -317,6 +325,9 @@ class Parameter extends Variable
         // If there is a default value, store it and its type
         $default_node = $node->children['default'];
         if ($default_node !== null) {
+            if (preg_match('/^(_$|unused)/iD', $parameter_name)) {
+                $parameter->should_warn_if_provided = true;
+            }
             // Set the actual value of the default
             $parameter->setDefaultValue($default_node);
             try {
@@ -784,5 +795,22 @@ class Parameter extends Variable
         if ($other->default_value_from_reflection) {
             $this->default_value_from_reflection = true;
         }
+    }
+
+    /**
+     * Sets whether phan should warn if this parameter is provided
+     * @suppress PhanUnreferencedPublicMethod this may be set by phpdoc comments in the future.
+     */
+    public function setShouldWarnIfProvided(bool $should_warn_if_provided): void
+    {
+        $this->should_warn_if_provided = $this->hasDefaultValue() && $should_warn_if_provided;
+    }
+
+    /**
+     * Returns true if this should warn if the parameter is provided
+     */
+    public function shouldWarnIfProvided(): bool
+    {
+        return $this->should_warn_if_provided;
     }
 }